server/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
Pillow==10.0.1
numpy==1.26.2
opencv-python==4.9.0.80 
scikit-learn==1.2.2
scikit-image==0.19.3
torch==2.3.1 # PyTorch for SAM
torchvision==0.15.2
diffusers==0.29.0
transformers==4.42.3
accelerate==0.31.0
matplotlib==3.8.1
tensorboardX==2.6.2

